# openapi.yaml
...
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  "/faculty/{id}":
    get:
      tags:
        - user
      summary: Gets faculty
      operationId: getFaculty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Faculty"
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                type: string
    put:
      tags:
        - faculty-controller
      operationId: updateFaculty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Faculty"
        required: true
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                type: string
    delete:
      tags:
        - faculty-controller
      operationId: deleteFaculty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                type: string
...
  "/faculty/add":
    post:
      tags:
        - faculty-controller
      operationId: addFaculty
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Faculty"
        required: true
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
          content:
            "*/*":
              schema:
                type: string
...
components:
  schemas:
    Faculty:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...